% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_coordinate}
\alias{check_coordinate}
\title{Check coordinate for errors}
\usage{
check_coordinate(coordinate)
}
\arguments{
\item{coordinate}{coordinate of the data point}
}
\value{
TRUE
}
\description{
Coordinate must satisfy: String.	Concatenation of the member ID values for
each dimension. (refer to the TableMetadata datastructure for information
about members). Maximum of 10 dimensions - a fixed length.
One value per dimension (i.e. 1.1.1.36.1.0.0.0.0.0)
}
\examples{
check_coordinate("1.1.1.36.1.0.0.0.0.0")

}
