% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_periods}
\alias{check_periods}
\title{Check number of periods}
\usage{
check_periods(periods)
}
\arguments{
\item{periods}{Number of periods}
}
\value{
TRUE
}
\description{
Periods must be a strictly positive integer.
}
\examples{
check_periods(10)

}
