% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_product_id}
\alias{check_product_id}
\title{Check product ID for errors}
\usage{
check_product_id(product_id)
}
\arguments{
\item{product_id}{ID of the product/table}
}
\value{
TRUE
}
\description{
Product ID must satisfy: 10-digit identifier (i.e. 1310008901).
First two digits are the Subject (i.e. 13=Health). Digits 3 and 4 are the
product type (i.e. 10=table/cube). Digits 5 to 8 are the sequential numbers
within the subject (i.e. 0089). Digits 9 and 10 are optional, to identify
the simple view of a table/cube (i.e. 01)
}
\examples{
\dontrun{
check_product_id(1310008901)
}
}
