% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_vector_id}
\alias{check_vector_id}
\title{Check vector ID for errors}
\usage{
check_vector_id(vector_id)
}
\arguments{
\item{vector_id}{ID of the Vector that represents the time series}
}
\value{
TRUE
}
\description{
Vector ID must satisfy: Vector number (i.e. 42973393).
Current max length: 10 digits Minimum length: 1.
}
\examples{
\dontrun{
check_vector_id(42973393)
}

}
