% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_vector}
\alias{extract_vector}
\title{Extract vector data from httr content response}
\usage{
extract_vector(content_vector)
}
\arguments{
\item{content_vector}{httr content response from getBulkVectorDataByRange call}
}
\value{
data.frame with 3 columns: vector_id, ref_date, value
}
\description{
Extract vector data from httr content response and return as data.frame
}
\examples{
\dontrun{
response <- getBulkVectorDataByRange("v113411623")
vector_content <- httr::content(response)
extract_vector(vector_content)
}

}
