% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_wds_json}
\alias{format_wds_json}
\title{Format json for API call}
\usage{
format_wds_json(...)
}
\arguments{
\item{...}{List of parameters}
}
\value{
JSON
}
\description{
Format json for API call
}
\examples{
\dontrun{

}

}
