% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-series-metadata.R
\name{getAllCubesList}
\alias{getAllCubesList}
\title{Get list of all available data tables}
\usage{
getAllCubesList()
}
\value{
A json object.
}
\description{
Users can query the output database to provide a complete inventory of
data tables available through this Statistics Canada API. This command
accesses a comprehensive list of details about each table including
information at the dimension level.
}
\examples{
\dontrun{
getAllCubesList()
}
}
