% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getBulkVectorDataByRange}
\alias{getBulkVectorDataByRange}
\title{Download a vector}
\usage{
getBulkVectorDataByRange(vectorIds)
}
\arguments{
\item{vectorIds}{Vector is a short identifier to refer to a time series of
data points. Unique variable length reference code, consisting of up to 10
digits. (i.e. 1234567890, 1, etc.)}
}
\value{
An httr response object
}
\description{
Download a vector
}
\examples{
\dontrun{
getBulkVectorDataByRange(74804)
getBulkVectorDataByRange(c(74804, 1))
}
}
