% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getChangedSeriesDataFromCubePidCoord}
\alias{getChangedSeriesDataFromCubePidCoord}
\title{Get changed series data}
\usage{
getChangedSeriesDataFromCubePidCoord(product_id, coordinate)
}
\arguments{
\item{product_id}{Product Identification number (PID) is a unique product
identifier for all Statistics Canada products, including large
multidimensional tables. The first two digits refer to a subject, the next
two digits refer to product type, the last four digits refer to the product
itself.}

\item{coordinate}{Coordinate is a concatenation of the member ID values for
each dimension. One value per dimension. (i.e. "1.3.1.1.1.1.0.0.0.0" ) A
table PID number combined with a coordinate will identify a unique time
series of data points.}
}
\value{
A json object.
}
\description{
Get changed series data given a product ID and coordinate
}
\examples{
\donttest{
getChangedSeriesDataFromCubePidCoord(35100003, "1.12.0.0.0.0.0.0.0.0")
}

}
