% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getChangedSeriesDataFromVector}
\alias{getChangedSeriesDataFromVector}
\title{Get changed series data from vector}
\usage{
getChangedSeriesDataFromVector(vector_id)
}
\arguments{
\item{vector_id}{Vector is a short identifier to refer to a time series of
data points. Unique variable length reference code, consisting of the
letter 'V', followed by up to 10 digits. (i.e. V1234567890, V1, etc.)}
}
\value{
A json object.
}
\description{
Get changed series data from vector
}
\examples{
\donttest{
getChangedSeriesDataFromVector("74804")
getChangedSeriesDataFromVector("v74804")
getChangedSeriesDataFromVector(74804)
}
}
