% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-series-metadata.R
\name{getCubeMetadata}
\alias{getCubeMetadata}
\title{Get cube/table metadata}
\usage{
getCubeMetadata(product_id)
}
\arguments{
\item{product_id}{Product Identification number (PID) is a unique product
identifier for all Statistics Canada products, including large
multidimensional tables. The first two digits refer to a subject, the next
two digits refer to product type, the last four digits refer to the product
itself.}
}
\value{
A json object.
}
\description{
Get cube/table titles, product ID, CANSIM ID, release date, and more.
}
\examples{
\donttest{
getCubeMetadata(35100003)
}
}
