% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getDataFromCubePidCoordAndLatestNPeriods}
\alias{getDataFromCubePidCoordAndLatestNPeriods}
\title{Get changed series data}
\usage{
getDataFromCubePidCoordAndLatestNPeriods(product_id, coordinate, periods)
}
\arguments{
\item{product_id}{Product Identification number (PID) is a unique product
identifier for all Statistics Canada products, including large
multidimensional tables. The first two digits refer to a subject, the next
two digits refer to product type, the last four digits refer to the product
itself.}

\item{coordinate}{Coordinate is a concatenation of the member ID values for
each dimension. One value per dimension. (i.e. "1.3.1.1.1.1.0.0.0.0" ) A
table PID number combined with a coordinate will identify a unique time
series of data points.}

\item{periods}{Number of periods to return, starting from the most recent
available data point.}
}
\value{
A json object.
}
\description{
Get changed series data given a product ID and coordinate
}
\examples{
\dontrun{
getDataFromCubePidCoordAndLatestNPeriods(35100003, "1.12.0.0.0.0.0.0.0.0", 10)
}

}
