% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getFullTableDownloadCSV}
\alias{getFullTableDownloadCSV}
\title{Get full table download link in CSV format}
\usage{
getFullTableDownloadCSV(product_id, language = c("en", "fr"))
}
\arguments{
\item{product_id}{Product Identification number (PID) is a unique product
identifier for all Statistics Canada products, including large
multidimensional tables. The first two digits refer to a subject, the next
two digits refer to product type, the last four digits refer to the product
itself.}

\item{language}{'en' for English or 'fr' for French}
}
\value{
A json object.
}
\description{
A direct link to download the table in CSV format. This functiom
does not actually download the table itself.
}
\examples{
\dontrun{
getFullTableDownloadCSV(35100003, "fr")
}
}
