% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getFullTableDownloadSDMX}
\alias{getFullTableDownloadSDMX}
\title{Get full table download link in SDMX format}
\usage{
getFullTableDownloadSDMX(product_id)
}
\arguments{
\item{product_id}{Product Identification number (PID) is a unique product
identifier for all Statistics Canada products, including large
multidimensional tables. The first two digits refer to a subject, the next
two digits refer to product type, the last four digits refer to the product
itself.}
}
\value{
A json object.
}
\description{
Get cube/table titles, product ID, CANSIM ID, release date, and more.
}
\examples{
\dontrun{
getFullTableDownloadSDMX(35100003)
}
}
