% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-series-metadata.R
\name{getSeriesInfoFromVector}
\alias{getSeriesInfoFromVector}
\title{Get series metadata from vector ID}
\usage{
getSeriesInfoFromVector(vector_id)
}
\arguments{
\item{vector_id}{Vector is a short identifier to refer to a time series of
data points. Unique variable length reference code, consisting of the
letter 'V', followed by up to 10 digits. (i.e. V1234567890, V1, etc.)}
}
\value{
An httr response object
}
\description{
Get series titles, product ID, CANSIM ID, release date, and more.
}
\examples{
\dontrun{
getSeriesInfoFromVector(32164132)
}

}
