% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stc_time}
\alias{stc_time}
\title{Format date for API call}
\usage{
stc_time(date)
}
\arguments{
\item{date}{A date}
}
\value{
Date with format \%Y-\%m-\%dT\%H:\%M
}
\description{
Format date to \%Y-\%m-\%dT\%H:\%M, using Statcan's time zone
}
\examples{
\dontrun{
check_vector_id(42973393)
}

}
