% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{getBulkVectorDataByRange}
\alias{getBulkVectorDataByRange}
\title{Download a vector}
\usage{
getBulkVectorDataByRange(
  vector_ids,
  start_release_date = "1901-01-01",
  end_release_date = as.Date(Sys.time())
)
}
\arguments{
\item{vector_ids}{Vector is a short identifier to refer to a time series of
data points. Unique variable length reference code, consisting of the
letter 'V', followed by up to 10 digits. (i.e. V1234567890, V1, etc.)}

\item{end_release_date}{}
}
\value{
A json object.
}
\description{
Download a vector
}
\examples{
\donttest{
getBulkVectorDataByRange("74804")
getBulkVectorDataByRange("v74804")
getBulkVectorDataByRange(74804)
}
}
