% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-data-service.R
\name{getBulkVectorDataByRange}
\alias{getBulkVectorDataByRange}
\title{Download a vector}
\usage{
getBulkVectorDataByRange(
  vectors,
  start_release_date = "1901-01-01",
  end_release_date = as.Date(Sys.time())
)
}
\arguments{
\item{end_release_date}{}
}
\value{
A json object.
}
\description{
Download a vector
}
\examples{
\donttest{
getBulkVectorDataByRange("74804")
}
}
