% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental-information.R
\name{getCodeSets}
\alias{getCodeSets}
\title{Get code sets}
\usage{
getCodeSets()
}
\value{
An httr response object
}
\description{
Code Sets provide additional information to describe the information such as
scales, frequencies and symbols. Use method getCodeSets to access the most
recent version of the code sets with descriptions (English and French) for
all possible codes.
}
\examples{
\dontrun{
getCodeSets()
}

}
