% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post}
\alias{post}
\title{POST utility}
\usage{
post(url_func, ...)
}
\arguments{
\item{url_func}{name of API function to be appended to wds_url}

\item{...}{body of POST request, as named arguments or a dataframe}
}
\value{
httr response object
}
\description{
POST utility
}
